backend/requirements.txt
fastapi==0.135.1
uvicorn[standard]==0.42.0
python-dotenv==1.2.2
pydantic==2.12.5

# LangGraph + LLM
langgraph==1.1.3
langchain-core==1.2.20
langchain-groq==1.1.2

# Vector DB + embeddings
qdrant-client==1.17.1
sentence-transformers==5.3.0

# Evaluation + experiment
numpy==2.4.3
matplotlib==3.10.8
tqdm==4.67.3

pypdf==6.9.1
